test: split test case for string/json
